+++ assembly/util/index.ts
@@ -3,3 +3,4 @@ export * from "./atoi";
 export * from "./concat";
 export * from "./bytes";
 export * from "./nextPowerOf2";
+export * from "./stringScan";